+++ app/pug/action_list.pug
@@ -6,17 +6,17 @@ block vars
 
 block content
   .container-fluid
-    div.vert-space-x2
+    .vert-space-x2
       if(singleType && actionTypeForm.tags && !actionTypeForm.tags.includes('Trash'))
-        a(href = `/action/${actionTypeForm.formId}`).btn.btn-primary
+        a.btn.btn-primary(href = `/action/${actionTypeForm.formId}`)
           img.small-icon(src = '/images/run_icon.svg')
           | &nbsp; Perform a New Action of this Type
 
-        div.vert-space-x2
+        .vert-space-x2
 
       h2 #{title}
 
-    div.vert-space-x1
+    .vert-space-x1
       table.table
         thead
           tr
@@ -44,7 +44,7 @@ block content
                 td
                   +dateFormat(action.lastEditDate)
                 td
-                  a(href = `/json/action/${action.actionId}`, target = '_blank').btn-sm.btn-secondary
+                  a.btn-sm.btn-secondary(href = `/json/action/${action.actionId}`, target = '_blank')
                     img.small-icon(src = '/images/checklist_icon.svg')
                     | &nbsp; JSON
-    div.vert-space-x2
+    .vert-space-x2
