+++ app/pug/action_specComponent.pug
@@ -15,10 +15,10 @@ block extrascripts
 
 block content
   .container-fluid
-    div.vert-space-x2
+    .vert-space-x2
       h2 Perform Action (Specified Component)
 
-    div.vert-space-x1
+    .vert-space-x1
       dl
         dt Action Type
         dd #{actionTypeForm.formName}
@@ -27,10 +27,10 @@ block content
         dd
           a(href = `/component/${componentUuid}`) #{componentUuid}
 
-    div.vert-space-x1
+    .vert-space-x1
       hr
 
-    div.vert-space-x1.formbuilder_area
+    .vert-space-x1.formbuilder_area
       #typeform
 
-    div.vert-space-x2
+    .vert-space-x2
