+++ app/pug/action_unspecComponent.pug
@@ -12,15 +12,15 @@ block extrascripts
 
 block content
   .container-fluid
-    div.vert-space-x2
+    .vert-space-x2
       h2 Perform Action (Unspecified Component)
 
-    div.vert-space-x1
+    .vert-space-x1
       dl
         dt Action Type
         dd #{actionTypeFormName}
 
-    div.vert-space-x1
+    .vert-space-x1
       hr
 
       p An action must be performed on a single specified component. Please enter the component UUID in the input box below.
@@ -29,11 +29,11 @@ block content
         br
         | Alternatively, click on the camera icon (to the right of the box) to scan the component's QR code if you have it available.
 
-    div.vert-space-x1
+    .vert-space-x1
       .row
         .col
           #uuidform
 
         .col.offset-md-20
 
-    div.vert-space-x2
+    .vert-space-x2
