+++ app/pug/component.pug
@@ -18,10 +18,10 @@ block extrascripts
 
 block content
   .container-fluid
-    div.vert-space-x2
+    .vert-space-x2
       h2 View Component Record
 
-    div.vert-space-x1
+    .vert-space-x1
       .row
         .col
           dl
@@ -33,7 +33,7 @@ block content
 
             dt Component UUID
             dd
-              div.form-inline
+              .form-inline
                 a(href = `/component/${component.componentUuid}`) #{component.componentUuid}
 
                 p &nbsp; &nbsp;
@@ -56,11 +56,11 @@ block content
                 |  by !{' '}
                 a(href = `mailto:${user_email(component.insertion.user)}`) #{component.insertion.user.displayName}
 
-              div.vert-space-x1.border-success.border.rounded.p-2
+              .vert-space-x1.border-success.border.rounded.p-2
                 #typeform
 
             if component.data.subComponent_fullUuids
-              div.vert-space-x1.border-success.border.rounded.p-2
+              .vert-space-x1.border-success.border.rounded.p-2
                 dt Sub Components
 
                 dd
@@ -91,13 +91,13 @@ block content
               dd
                 a(href = `/component/${component.data.fromBatch}`, target = '_blank') #{component.data.fromBatch}
 
-          div.vert-space-x1
+          .vert-space-x1
             .panel.panel-default
             .foldable-heading.panel-heading.collapsed(data-toggle = 'collapse' data-target = '#versionsBox')
               | All Versions: 
               i.chevron.fa.fa-fw 
 
-            div#versionsBox.collapse
+            .collapse#versionsBox
               ul
                 each v in componentVersions
                   li 
@@ -110,33 +110,33 @@ block content
                       a(href = `mailto:${user_email(v.insertion.user)}`) #{v.insertion.user.displayName}
 
         .col-md-3
-          div#qr-code
+          #qr-code
             +qr-panel(`${base_url}/c/${component.shortUuid.toString()}`, component.data.name)
 
-    div.vert-space-x2.form-inline
-      a(href = `/component/${component.componentUuid}/edit`).btn.btn-warning
+    .vert-space-x2.form-inline
+      a.btn.btn-warning(href = `/component/${component.componentUuid}/edit`)
         img.small-icon(src = '/images/edit_icon.svg')
         | &nbsp; Edit Component
 
-      div.hori-space-x2
-        a(href = `/json/component/${component.componentUuid}`, target = '_blank').btn.btn-secondary
+      .hori-space-x2
+        a.btn.btn-secondary(href = `/json/component/${component.componentUuid}`, target = '_blank')
           img.small-icon(src = '/images/checklist_icon.svg')
           | &nbsp; View JSON Record
 
-      div.hori-space-x2
-        a(href = `/component/${component.componentUuid}/qrCodes`).btn.btn-primary
+      .hori-space-x2
+        a.btn.btn-primary(href = `/component/${component.componentUuid}/qrCodes`)
           img.small-icon(src = '/images/checklist_icon.svg')
           | &nbsp; Print a set of the component's QR codes
 
-      div.hori-space-x2
-        a(href = `/component/${component.componentUuid}/summary`).btn.btn-primary
+      .hori-space-x2
+        a.btn.btn-primary(href = `/component/${component.componentUuid}/summary`)
           img.small-icon(src = '/images/checklist_icon.svg')
           | &nbsp; View and print the component's summary
 
-    div.vert-space-x2
+    .vert-space-x2
       hr
 
-    div.vert-space-x1
+    .vert-space-x1
       .row
         .col-sm-6
           h4 Action History
@@ -176,7 +176,7 @@ block content
 
           each typeForm in actionTypeForms
             if(typeForm.componentTypes.includes(component.formName) && !typeForm.tags.includes('Trash'))
-              a(href = `/action/${typeForm.formId}/${component.componentUuid}`).tagbutton.btn.btn-success #{typeForm.formName}
+              a.actionbutton.btn.btn-success(href = `/action/${typeForm.formId}/${component.componentUuid}`) #{typeForm.formName}
 
           hr
 
@@ -184,6 +184,6 @@ block content
 
           each typeForm in actionTypeForms
             if(!typeForm.componentTypes.includes(component.formName) && !typeForm.tags.includes('Trash'))
-              a(href = `/action/${typeForm.formId}/${component.componentUuid}`).tagbutton.btn.btn-success #{typeForm.formName}
+              a.actionbutton.btn.btn-success(href = `/action/${typeForm.formId}/${component.componentUuid}`) #{typeForm.formName}
 
-    div.vert-space-x2
+    .vert-space-x2
