+++ app/pug/component_edit.pug
@@ -17,10 +17,10 @@ block extrascripts
 
 block content
   .container-fluid
-    div.vert-space-x2
+    .vert-space-x2
       h2 Create or Edit Component
 
-    div.vert-space-x1
+    .vert-space-x1
       dl
         dt Component Type
         dd #{componentTypeForm.formName}
@@ -28,10 +28,10 @@ block content
         dt Component UUID
         dd #{component.componentUuid}
 
-    div.vert-space-x1
+    .vert-space-x1
       hr
 
-    div.vert-space-x1.formbuilder_area
+    .vert-space-x1.formbuilder_area
       #typeform
 
-    div.vert-space-x2
+    .vert-space-x2
