+++ app/pug/home.pug
@@ -35,7 +35,7 @@ block content
       p
 
   .container-fluid
-    div.vert-space-x1.border-success.border.rounded.p-3
+    .vert-space-x1.border-success.border.rounded.p-3
       h5 Useful Links, Bugs and Other Issues
       ul
         li For documentation relating to the database itself (i.e. descriptions of the functionality, overviews of common usage, etc.), please see our 
@@ -54,4 +54,4 @@ block content
           |  (a Github account is required to report issues).  Please give as much information as you can - an example page or link to where the bug occurs, how to reproduce the error (if appropriate), etc.
         li Requests for new and/or additional features and functionality can also be made on the Issues page, but please first check if the addition and/or change can be made by users directly through the web interface.
 
-    div.vert-space-x2
+    .vert-space-x2
