+++ app/pug/search_byLocationOrPartNumber.pug
@@ -9,17 +9,17 @@ block extrascripts
 
 block content
   .container-fluid
-    div.vert-space-x2
+    .vert-space-x2
       h2 Search for Geometry Boards by Location or Part Number
 
-    div.vert-space-x2
+    .vert-space-x2
       hr
 
       .row
         .col-md-4
           h4 Select Location
 
-          div.vert-space-x1
+          .vert-space-x1
             p Select a location from the options below.
 
           select.form-control#locationSelection
@@ -35,12 +35,12 @@ block content
             option(value = 'uwPsl') UW / PSL 
             option(value = 'chicago') Chicago
 
-          div.vert-space-x2
+          .vert-space-x2
             hr
 
             h4 Select Part Number
 
-            div.vert-space-x1
+            .vert-space-x1
               p Select a part number from the options below.
 
             select.form-control#partNumberSelection
@@ -80,7 +80,7 @@ block content
         .col-md-7
           h4 Search Results
 
-          div.vert-space-x1
+          .vert-space-x1
             table#results
 
-    div.vert-space-x2
+    .vert-space-x2
