+++ app/pug/search_byUUIDOrID.pug
@@ -9,48 +9,48 @@ block extrascripts
 
 block content
   .container-fluid
-    div.vert-space-x2
+    .vert-space-x2
       h2 Search for Record by UUID or ID
 
-    div.vert-space-x2 
+    .vert-space-x2 
       hr
 
       h4 Search for Component by UUID
 
-      div.vert-space-x1
+      .vert-space-x1
         p The input box below uses auto-complete - just start typing a component UUID, and any matching options will be displayed.
           br
           | Alternatively, click on the camera icon (to the right of the box) to take a picture of the component's QR code if you have it available.
 
-      div.vert-space-x1
+      .vert-space-x1
         .row
           .col-md-6
             #componentuuidform
 
-    div.vert-space-x2
+    .vert-space-x2
       hr
 
       h4 Search for Action by ID
 
-      div.vert-space-x1
+      .vert-space-x1
         p The input box below uses auto-complete - just start typing an action ID, and any matching options will be displayed.
 
-      div.vert-space-x1
+      .vert-space-x1
       .row
         .col-md-6
           #actionidform
 
-    div.vert-space-x2
+    .vert-space-x2
       hr
 
       h4 Search for Workflow by ID
 
-      div.vert-space-x1
+      .vert-space-x1
         p The input box below uses auto-complete - just start typing a workflow ID, and any matching options will be displayed.
 
-      div.vert-space-x1
+      .vert-space-x1
       .row
         .col-md-6
           #workflowidform
 
-    div.vert-space-x2
+    .vert-space-x2
