+++ app/pug/search_byVisInspectOrOrderNumber.pug
@@ -9,17 +9,17 @@ block extrascripts
 
 block content
   .container-fluid
-    div.vert-space-x2
+    .vert-space-x2
       h2 Search for Geometry Boards by Visual Inspection Disposition
 
-    div.vert-space-x2 
+    .vert-space-x2 
       hr
 
       .row
         .col-md-4
           h4 Select Disposition
 
-          div.vert-space-x1
+          .vert-space-x1
             p Select a visual inspection disposition from the options below.
 
           select.form-control#dispositionSelection
@@ -36,17 +36,17 @@ block content
         .col-md-5
           h4 Enter Order Number
 
-          div.vert-space-x1
+          .vert-space-x1
             p Enter a board order number below and press the 'Enter' key.
 
           input.form-control#orderNumberSelection
 
-    div.vert-space-x2
+    .vert-space-x2
       hr
 
       h4 Search Results
 
-      div.vert-space-x1
+      .vert-space-x1
         table#results
 
-    div.vert-space-x2
+    .vert-space-x2
