+++ app/pug/user_list.pug
@@ -10,10 +10,10 @@ block extrascripts
 
 block content
   .container-fluid
-    div.vert-space-x2
+    .vert-space-x2
       h2 Users
 
-    div.vert-space-x1
+    .vert-space-x1
       table.table#users_table
         thead
           tr
@@ -24,4 +24,4 @@ block content
 
         tbody
 
-    div.vert-space-x2
+    .vert-space-x2
