+++ app/pug/workflow_edit.pug
@@ -14,18 +14,18 @@ block extrascripts
 
 block content
   .container-fluid
-    div.vert-space-x2
+    .vert-space-x2
       h2 Create or Edit Workflow
 
-    div.vert-space-x1
+    .vert-space-x1
       dl
         dt Workflow Type
         dd #{workflowTypeForm.formName}
 
-    div.vert-space-x1
+    .vert-space-x1
       hr
 
-    div.vert-space-x1.formbuilder_area
+    .vert-space-x1.formbuilder_area
       #typeform
 
-    div.vert-space-x2
+    .vert-space-x2
