+++ app/pug/workflow_list.pug
@@ -6,17 +6,17 @@ block vars
 
 block content
   .container-fluid
-    div.vert-space-x2
+    .vert-space-x2
       if(singleType && workflowTypeForm.tags && !workflowTypeForm.tags.includes('Trash'))
-        a(href = `/workflow/${workflowTypeForm.formId}`).btn.btn-primary
+        a.btn.btn-primary(href = `/workflow/${workflowTypeForm.formId}`)
           img.small-icon(src = '/images/run_icon.svg')
           | &nbsp; Create a New Workflow of this Type
 
-        div.vert-space-x2
+        .vert-space-x2
 
       h2 #{title}
 
-    div.vert-space-x1
+    .vert-space-x1
       table.table
         thead
           tr
@@ -50,8 +50,8 @@ block content
                   td.text-danger.font-weight-bold #{workflow.status}
 
                 td
-                  a(href = `/json/workflow/${workflow.workflowId}`, target = '_blank').btn-sm.btn-secondary
+                  a.btn-sm.btn-secondary(href = `/json/workflow/${workflow.workflowId}`, target = '_blank')
                     img.small-icon(src = '/images/checklist_icon.svg')
                     | &nbsp; JSON
 
-    div.vert-space-x2
+    .vert-space-x2
