+++ app/routes/workflows.js
@@ -1,12 +1,9 @@
-const { readFileSync } = require('fs');
 const router = require('express').Router();
 
 const Forms = require('lib/Forms.js');
 const logger = require('../lib/logger');
 const permissions = require('lib/permissions.js');
 const Workflows = require('lib/Workflows.js');
-
-const default_form_schema = JSON.parse(readFileSync('./schemas/default_form_schema.json'));
 
 
 /// View a single workflow record
@@ -167,14 +164,14 @@ router.get('/workflowTypes/:typeFormId/new', permissions.checkPermission('forms:
     // Attempt to retrieve any and all existing type forms with this type form ID
     let typeForm = await Forms.retrieve('workflowForms', req.params.typeFormId);
 
-    // If there are no existing type forms, set up a new one using the specified type form ID and the default form schema
+    // If there are no existing type forms, set up a new one using the specified type form ID and an initially empty form schema
     // Then save the new type form into the 'actionForms' collection of records
     // Initially, use the form ID as the form name as well - the user will have the option of changing the name later
     if (!typeForm) {
       typeForm = {
         formId: req.params.typeFormId,
         formName: req.params.typeFormId,
-        schema: default_form_schema,
+        schema: { components: [] },
       };
 
       Forms.save(typeForm, 'workflowForms', req);
