--- app/schemas/default_form_schema.json
@@ -1,4 +1,0 @@
-{
-  "components": [
-  ]
-}
