+++ app/static/css/component.css
@@ -9,6 +9,6 @@
   border: 2px solid #008CBA;
 }
 
-.tagbutton {
+.actionbutton {
   margin: 4px;
 }
