+++ app/static/js/qrcodegen.js
