+++ app/static/pages/component.js
@@ -2,7 +2,7 @@
 window.addEventListener('load', populateTypeForm);
 
 
-/// Function to run when the page is loaded
+// Function to run when the page is loaded
 async function populateTypeForm() {
   // Render the component type form in the page element called 'typeform'
   let typeForm = await Formio.createForm(document.getElementById('typeform'), componentTypeForm.schema, { readOnly: true, });
