+++ app/static/pages/component_edit.js
@@ -19,7 +19,7 @@ if (subComponent_fullUuids.length > 0) newBatchComponent = true;
 window.addEventListener('load', onPageLoad);
 
 
-/// Function to run when the page is loaded
+// Function to run when the page is loaded
 async function onPageLoad() {
   // Set up a new type form based on the schema of the component type form
   let schema = componentTypeForm.schema;
@@ -153,7 +153,7 @@ async function onPageLoad() {
 }
 
 
-/// Function to submit the completed 'submission' object to the database
+// Function to submit the completed 'submission' object to the database
 function SubmitData(submission, redirectToInfo = true) {
   // Submit the 'submission' object via a jQuery 'ajax' call, with the success and failure functions as defined below
   $.ajax({
@@ -166,7 +166,7 @@ function SubmitData(submission, redirectToInfo = true) {
   }).fail(postFail);
 
 
-  /// Function to run for a successful submission
+  // Function to run for a successful submission
   function postSuccess(result) {
     // If the submission result contains an error (even with a successful submission), display it along with the appropriate Formio alert type
     if (result.error) {
@@ -190,7 +190,7 @@ function SubmitData(submission, redirectToInfo = true) {
   }
 
 
-  /// Function to run for a failed submission
+  // Function to run for a failed submission
   function postFail(result, statusCode, statusMsg) {
     // If the submission result contains a response message, display it along with the appropriate Formio alert type
     // Otherwise, display any status message and error code instead
